+++ CHANGELOG.md
@@ -312,6 +312,7 @@ The major version bump is due to dropped support for PHP `7.2` and a several bre
 * The Gatekeeper `Adapter::activateUser()` method will now always return a boolean value as described in the documentation.
 * Renamed the `InputValidationTrait::validate()` method to `InputValidationTrait::getValidatedInput()`.
 * Removed the undocumented recursive configuration file merging.
+* Renamed `Validator::validate()` to `Validator::getValidatedInput()`.
 
 #### Improvements